metricsの移植成功. PATHも問題なし
